LayoutTests/imported/w3c/web-platform-tests/preload/preload-referrer-policy-subresource-header.tentative_isCrossOriginPreload=false&isCrossOriginResource=true-expected.txt

The referrerpolicy attribute on Link header should be ignored for subresources to prevent cross-origin referrer leakage


FAIL referrer policy ( -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy ( -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
PASS referrer policy (no-referrer -> , same-origin, cross-origin)
PASS referrer policy (no-referrer -> no-referrer, same-origin, cross-origin)
PASS referrer policy (no-referrer -> same-origin, same-origin, cross-origin)
PASS referrer policy (no-referrer -> origin, same-origin, cross-origin)
PASS referrer policy (no-referrer -> origin-when-cross-origin, same-origin, cross-origin)
PASS referrer policy (no-referrer -> strict-origin-when-cross-origin, same-origin, cross-origin)
PASS referrer policy (no-referrer -> unsafe-url, same-origin, cross-origin)
FAIL referrer policy (same-origin -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (same-origin -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/"
FAIL referrer policy (origin-when-cross-origin -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (origin-when-cross-origin -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (strict-origin-when-cross-origin -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> , same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> no-referrer, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> same-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> strict-origin-when-cross-origin, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
FAIL referrer policy (unsafe-url -> unsafe-url, same-origin, cross-origin) assert_equals: expected "NO-REFERER" but got "http://web-platform.test:8800/preload/preload-referrer-policy-subresource-header.tentative.html?isCrossOriginPreload=false&isCrossOriginResource=true"
